requirements-gpu.txt

annotated-types==0.7.0
anyio==3.7.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fastapi==0.104.1
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
h11==0.16.0
huggingface-hub==0.32.3
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==1.26.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pip==25.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.3
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
thop==0.1.1-2209072238
tokenizers==0.15.2
torch==2.1.1
torchvision==0.16.1
tqdm==4.67.1
transformers==4.35.2
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
ultralytics==8.3.147
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.24.0
wheel==0.45.1
